Organize API and proxy code into smaller, focused files
Restructure internal API and proxy logic into multiple smaller files for improved maintainability and readability.
